Seed recordings on the fly instead of using code generation
