+++ e2e/.gitignore
@@ -4,7 +4,6 @@ playwright-report/
 blob-reports/
 playwright/.cache/
 runner/e2e
-runner/gen-recording-events
 runner/runner
 certs/
 data/
