+++ e2e/runner/main.go
@@ -17,7 +17,6 @@
  */
 
 //go:generate go run ./cmd/gen-ts-fixtures
-//go:generate go run ./cmd/gen-recording-events
 
 package main
 
@@ -37,7 +36,6 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"github.com/gravitational/teleport/e2e/runner/fixtures"
-	"github.com/gravitational/teleport/e2e/runner/testenv"
 )
 
 var logLevel = new(slog.LevelVar)
@@ -289,7 +287,7 @@ func run(flags *e2eFlags, mode runMode, e2eDir string, isCI bool) error {
 		for _, inst := range allInstances {
 			outPath := filepath.Join(e2eDir, "config", inst.browser+"-teleport.yaml")
 			tcfg, err := generateTeleportConfig(config.teleportConfigTemplate, outPath, &TeleportConfig{
-				ClusterName:    testenv.ClusterName,
+				ClusterName:    clusterName,
 				DataDir:        inst.dataDir,
 				AuthServerPort: inst.authPort,
 				ProxyPort:      inst.proxyPort,
@@ -325,7 +323,7 @@ func run(flags *e2eFlags, mode runMode, e2eDir string, isCI bool) error {
 					return fmt.Errorf("failed to start Teleport for %s: %w", inst.browser, err)
 				}
 				if err := teleport.waitReady(gctx, 30*time.Second); err != nil {
-					return fmt.Errorf("Teleport for %s failed to become ready: %w", inst.browser, err)
+					return fmt.Errorf("teleport for %s failed to become ready: %w", inst.browser, err)
 				}
 				if err := seedRecordings(gctx, config.e2eDir, inst.dataDir); err != nil {
 					return fmt.Errorf("failed to seed session recordings for %s: %w", inst.browser, err)
