--- e2e/testdata/recordings/events.jsonl
@@ -1,1 +1,0 @@
-{"ei":23,"event":"session.end","uid":"54c8723c-e91e-4c34-affe-7dddeeda949a","code":"T2004I","time":"2026-03-27T16:36:19.938Z","cluster_name":"teleport-e2e","user":"bob","login":"root","user_kind":1,"user_roles":["access","admin","auditor","editor","updater_admin","awsidentitycenter"],"user_cluster_name":"teleport-e2e","sid":"e3286c06-d0ec-4d84-aeab-75692b885e3b","private_key_policy":"none","addr.remote":"[::1]:61152","proto":"ssh","namespace":"default","server_id":"59f4906c-d997-4b77-b5cb-0b29346f6469","server_hostname":"im-a-nodename","server_addr":"[::]:3022","server_labels":{"env":"example","hostname":"Yassines-MacBook-Pro.local"},"server_sub_kind":"teleport","server_version":"19.0.0-prealpha.2","enhanced_recording":false,"interactive":true,"participants":["bob"],"session_start":"2026-03-27T16:36:15.040647Z","session_stop":"2026-03-27T16:36:19.938249Z","session_recording":"node"}
